fix bug where season is not provided for inactive players so defaults to current
