+++ test/test_jockbot_nhl.py
@@ -3,6 +3,7 @@ import unittest
 import types
 
 from jockbot_nhl import nhl
+from jockbot_nhl import _helpers
 
 
 class TestNHL(unittest.TestCase):
@@ -13,9 +14,11 @@ class TestNHL(unittest.TestCase):
         self.league = nhl.NHL()
         self.team = nhl.NHLTeam(self.team_city)
         self.player = 'patrice bergeron'
+        self.inactive_player = 'wayne gretzky'
+        self.inactive_player_id = 8447400
         self.team_id = 6
         self.player_id = 8470638
-        self.config = self._get_config()
+        self.config = _helpers._get_config()
         self.stat_keys = self.config['stat_keys']
         self.goalie_stat = 'goalsAgainstAverage'
         self.skater_stat = 'points'
@@ -36,24 +39,26 @@ class TestNHL(unittest.TestCase):
         self.assertTrue(isinstance(self.league.wildcard_standings, dict), 'No Wildcard Standings')
 
     def test_team_id(self):
-        team_id = nhl._team_id(self.team_city)
+        team_id = _helpers._team_id(self.team_city)
         self.assertEqual(team_id, 6, 'Incorrect Team ID')
 
     def test_divisions(self):
         """Test _divisions function"""
-        divisions = nhl._divisions()
+        divisions = _helpers._divisions()
         message = f"divisions returned {type(divisions)} not generator"
         self.assertTrue(isinstance(divisions, types.GeneratorType), message)
 
     def test_games_on_date(self):
         """Test _games_on_date function"""
-        games = nhl._games_on_date('2019-05-27')
+        games = _helpers._games_on_date('2019-05-27')
         message = f"Games: {games}"
         self.assertTrue(isinstance(games, dict), message)
 
     def test_player_id(self):
-        player_id = nhl._player_id(self.player)
+        player_id = _helpers._player_id(self.player)
+        inactive_player_id = _helpers._player_id(self.inactive_player)
         self.assertEqual(player_id, self.player_id, 'Incorrect Player ID')
+        self.assertEqual(inactive_player_id, self.inactive_player_id, 'Incorrect Inactive Player ID')
 
     def test_wild_card_standings(self):
         """Test nhl._wild_card_standings"""
@@ -64,7 +69,7 @@ class TestNHL(unittest.TestCase):
     def test_fetch_league_leaders_skaters(self):
         """Test nhl._league_leaders function for skater stats"""
         stat_keys = self.stat_keys['skaters']
-        leaders = nhl._fetch_league_leaders('points', 'skater')
+        leaders = _helpers._fetch_league_leaders('points', 'skater')
         self.assertTrue(isinstance(leaders, list), 'Incorrect skater Type')
         self.assertEqual(len(leaders), 10, f"Incorrect number of players: {len(leaders)}")
         message = f"Skater Stat Keys Do Not Match\n{leaders[0].keys()}"
@@ -73,7 +78,7 @@ class TestNHL(unittest.TestCase):
     def test_fetch_league_leaders_goalies(self):
         """Test nhl._league_leaders function for goalie stats"""
         stat_keys = self.stat_keys['goalies']
-        leaders = nhl._fetch_league_leaders('saves', 'goalie')
+        leaders = _helpers._fetch_league_leaders('saves', 'goalie')
         message = f"Skater Stat Keys Do Not Match\n{leaders[0].keys()}"
         self.assertTrue(isinstance(leaders, list), 'Incorrect goalie Type')
         self.assertEqual(len(leaders), 10, f"Incorrect number of players: {len(leaders)}")
@@ -87,7 +92,7 @@ class TestNHL(unittest.TestCase):
         self.assertTrue(isinstance(leaders, dict), 'Incorrect leaders Type')
 
     def test_all_player_ids(self):
-        players = nhl._all_player_ids()
+        players = _helpers._all_player_ids()
         self.assertTrue(isinstance(players, dict), 'Incorrect players Type')
         self.assertEqual(players[self.player], self.player_id, 'Incorrect Player ID')
 
@@ -119,8 +124,10 @@ class TestNHL(unittest.TestCase):
 
     def test_get_player_stats(self):
         """Test NHL.get_player_stats function"""
-        stats = self.league.get_player_stats(player_name=self.player)
-        self.assertTrue(isinstance(stats, dict), 'No Player Stats')
+        player_stats = self.league.get_player_stats(player_name=self.player)
+        inactive_player_stats = self.league.get_player_stats(player_name=self.inactive_player, season='19871988')
+        self.assertTrue(isinstance(player_stats, dict), 'No Player Stats')
+        self.assertTrue(isinstance(inactive_player_stats, dict), 'No Inactive Player Stats')
 
     def test_get_career_stats(self):
         """Test NHL.get_career_stats function"""
